% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliertree.R
\name{as.list.outlieroutputs}
\alias{as.list.outlieroutputs}
\title{Convert outlier outputs to R list}
\usage{
\method{as.list}{outlieroutputs}(x, ...)
}
\arguments{
\item{x}{Outliers flagged by an \link{outlier.tree} model, returned by e.g. the
prediction function.}

\item{...}{Not used.}
}
\value{
The same outliers as a list of lists.
}
\description{
Converts outliers results (an object as returned from \link{predict.outliertree}
or from \link{extract.training.outliers}) to an R list which can be modified
programatically.
}
