% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliertree.R
\name{check.outlierness.bounds}
\alias{check.outlierness.bounds}
\title{Check values that could potentially flag an observation as outlier}
\usage{
check.outlierness.bounds(outlier_tree_model)
}
\arguments{
\item{outlier_tree_model}{An Outlier Tree model object as generated by `outlier.tree`.}
}
\value{
A list with column as the names and the bounds or categories as values.
}
\description{
Returns, for each numeric/date/timestamp column, a range of values *outside* of which
observations could potentially be flagged as being an outlier in some cluster, and
for categorical/ordinal/boolean columns, the factor levels that can be flagged as
being an outlier in some cluster. If the lower bound is higher than the upper bound, it means
any value can potentially be flagged as outlier.
}
