% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliertree.R
\name{extract.training.outliers}
\alias{extract.training.outliers}
\title{Extract outliers found in training data}
\usage{
extract.training.outliers(outlier_tree_model)
}
\arguments{
\item{outlier_tree_model}{An Outlier Tree object as returned by `outlier.tree`.}
}
\value{
A data frame with the outliers, which can be pretty-printed by function
`print` from this same package.
}
\description{
Extracts outliers from a model generated by `outlier.tree` if
it was passed parameter `save_outliers` = `TRUE`.
}
