% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothyroid.R
\docType{data}
\name{hypothyroid}
\alias{hypothyroid}
\title{Data about thyroid hormones for anonymous patients}
\format{An object of class \code{data.frame} with 2772 rows and 23 columns.}
\source{
\url{http://archive.ics.uci.edu/ml/datasets/thyroid+disease}
}
\usage{
data(hypothyroid)
}
\description{
This data contains several obvious outliers from misspellings in data entry.
From Garavan Institute. For more details see link at the bottom.
}
\keyword{datasets}
