% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obj_methods.R
\name{print.outliertree}
\alias{print.outliertree}
\title{Print summary information from Outlier Tree model}
\usage{
\method{print}{outliertree}(x, ...)
}
\arguments{
\item{x}{An Outlier Tree model as produced by function `outlier.tree`.}

\item{...}{Not used.}
}
\description{
Displays general statistics from a fitted Outlier Tree model (same as `summary`).
For printing the outliers discovered, use function `print` on the returned outliers
(e.g. from `predict`), not on the model object iself.
}
\details{
Note that after loading a serialized object from `outlier.tree` through `readRDS` or `load`,
it will only de-serialize the underlying C++ object upon running `predict` or `print`, so the first run will
be slower, while subsequent runs will be faster as the C++ object will already be in-memory.
}
