% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obj_methods.R
\name{summary.outliertree}
\alias{summary.outliertree}
\title{Print summary information from Outlier Tree model}
\usage{
\method{summary}{outliertree}(object, ...)
}
\arguments{
\item{object}{An Outlier Tree model as produced by function `outlier.tree`.}

\item{...}{Not used.}
}
\description{
Displays general statistics from a fitted Outlier Tree model (same as `print`).
For printing the outliers discovered, use function `print` on the returned outliers
(e.g. from `predict`), not on the model object iself.
}
