% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/titanic.R
\docType{data}
\name{titanic}
\alias{titanic}
\title{Data about passengers of the RMS Titanic}
\format{
An object of class \code{data.frame} with 1309 rows and 14 columns.
}
\source{
\url{https://github.com/jbryer/CompStats/raw/master/Data/titanic3.csv}
}
\usage{
data(titanic)
}
\description{
This dataset contains information about 1,309 of the passengers who boarded the
\href{https://en.wikipedia.org/wiki/Titanic}{RMS Titanic ship} for its last
trip in the year 1912 (spoiler: the ship doesn't reach its destination).
}
\keyword{datasets}
