% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obj_methods.R
\name{variable.names.outliertree}
\alias{variable.names.outliertree}
\title{Get Variable Names for OutlierTree Model}
\usage{
\method{variable.names}{outliertree}(object, ...)
}
\arguments{
\item{object}{An OutlierTree model as returned by function \link{outlier.tree}.}

\item{...}{Not used.}
}
\value{
A character vector with the variable names.
}
\description{
Returns the variable names from the data to which an OutlierTree model was fitted.

Columns will be returned in the following order according to their types, regardless of the order
that they had in theoriginal input data:\itemize{
\item{1.} Numeric.
\item{2.} Date.
\item{3.} Timestamp (POSIXct, POSIXlt).
\item{4.} Categorical.
\item{5.} Boolean / logical.
\item{6.} Ordinal.
}
}
